+++ idea-plugin/tests/com/reason/lang/napkin/SignatureParsingTest.java
@@ -136,4 +136,18 @@ public class SignatureParsingTest extends NsParsingTestCase {
   //    assertEquals("(string, payload, ~meta: 'meta=?, unit) => opaqueFsa",
   // signature.asString(myLanguage));
   // }
+
+  public void test_GH_275b() {
+    PsiSwitch e =
+        firstOfType(
+            parseCode(
+                "switch (a, b) { | (Some(a'), Some(b')) => let _ = { switch (x) { | None => None } } }"),
+            PsiSwitch.class);
+
+    assertEquals(
+        "switch (a, b) { | (Some(a'), Some(b')) => let _ = { switch (x) { | None => None } } }",
+        e.getText());
+    assertEquals(
+        "let _ = { switch (x) { | None => None } }", e.getPatterns().get(0).getBody().getText());
+  }
 }